rest_blog_django/requirements.txt
asgiref==3.2.7
Django==3.0.3
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-extensions==3.1.3
django-js-asset==1.2.2
django-recaptcha==2.0.6
djangorestframework==3.11.0
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
